Improve hero grid
This tweaks the hero grid system to have slightly different behaviour.
On desktop and above, the body content is 2/3 wide and the image 1/3. On
tablet, it's 1/2 each. On mobile, it stacks as usual.

In addition to that, this changes the behaviour of images; there's now
an "inline image" which is shown on mobile, and an "aside image" which
is shown on other viewports.