Fix layout and role of the <main> element
Internet Explorer < 11 does not style <main> elements as blocks, so we need to specify that in order for margin-bottom to have any effect in those browser.

Additionally, for older browsers using assistive technologies the main element does not get the correct ARIA role, so we’ll explicitly give it the main role to work around this.