Reconcile product page header with tech docs
We replaced the checkbox-based navigation toggle on the tech docs as although it meant Javascript wasn’t required for the toggle to work, it wasn’t as accessible as a button with the proper aria markup.

This brings the product page example in line. Unfortunately it does add a dependency on both jQuery and Modernizr (which were already required for the tech docs prior to these changes) but does mean that they are consistent in the way the navigation works.

This will hopefully allow us to extract the common styles and scripts both into a shared package of some description.