Add napa as a devDependency
This is so we can npm install govuk_template, as it doesn’t have a
package.json file.

We need the source scss files and these aren’t provided by packaged
versions of the template.

govuk_template is then installed to node_modules/ and we can refer to
the source scss files in core.scss.