Pin current latest govuk_template
Without any more specificity, napa will just install the latest
version of a repo's master branch into the local `/node_modules`.
So, without an explicit version, every time there is a commit to
govuk_template, we'll end up with slightly different code.

Pinning a version means we get the same thing every time, and the
new syntax in `package.json` lets us keep the folder name the same
as it was before.