fix: sort out the assets pipeline
1. move to `sassc`;

2. after a long investigation it seems that Sprockets needs a plain
CSS manifest (`application.css`) that can then use the Sprockets
directive (`link_tree/require`) to include and tie assets
together (merging the mere two lines of `application.css` at the top
of `screen.css.scss` wouldn't yield the right results).

Once Sprockets is correctly hooked up, it will effectively drag our
`govuk-frontend` resources into the assets folders (default:
`/assets`) so the last step is to indicate the path to find the
relevant resources to these stylesheets (namely for the fonts and
images). This is what the `$govuk-assets-path` variable lives for.

This also means we can retire from the custom `ImpostedAssetPathProcessor`
game which is not a fun one to be in.